personal-style-recommender/requirements.txt
tensorflow
requests
pandas
pillow
selenium
beautifulsoup4
webdriver-manager
